+++ app/login/action.ts
@@ -1,0 +1,25 @@
+'use server'
+
+import { revalidatePath } from 'next/cache'
+import { redirect } from 'next/navigation'
+import { createClient } from '@/utils/supabase/server'
+
+export async function login(formData: FormData) {
+  const supabase = await createClient()
+
+  const data = {
+    email: formData.get('email') as string,
+    password: formData.get('password') as string,
+  }
+
+  const { error } = await supabase.auth.signInWithPassword(data)
+
+  if (error) {
+    console.log(error.code)
+    redirect('/error')
+    
+  }
+
+  revalidatePath('/', 'layout')
+  redirect('/')
+}
