+++ app/logout/action.ts
@@ -1,0 +1,18 @@
+'use server'
+
+import { revalidatePath } from 'next/cache'
+import { redirect } from 'next/navigation'
+import { createClient } from '@/utils/supabase/server'
+
+export async function signup() {
+    const supabase = await createClient()
+  
+    const { error } = await supabase.auth.signOut()
+  
+    if (error) {
+      redirect('/error')
+    }
+  
+    revalidatePath('/', 'layout')
+    redirect('/login')
+  }
