+++ app/signup/action.ts
@@ -1,0 +1,26 @@
+'use server'
+
+import { revalidatePath } from 'next/cache'
+import { redirect } from 'next/navigation'
+import { createClient } from '@/utils/supabase/server'
+
+export async function signup(formData: FormData) {
+    const supabase = await createClient()
+  
+    // type-casting here for convenience
+    // in practice, you should validate your inputs
+    const data = {
+      email: formData.get('email') as string,
+      password: formData.get('password') as string,
+    }
+  
+    const { error } = await supabase.auth.signUp(data)
+  
+    if (error) {
+      console.log(error.code)
+      redirect('/error')
+    }
+  
+    revalidatePath('/', 'layout')
+    redirect('/')
+  }
